term calculation

+++ src/main/java/com/pmvaadin/calendars/dayofweeksettings/DefaultDaySetting.java
@@ -1,6 +1,5 @@
 package com.pmvaadin.calendars.dayofweeksettings;
 
-import java.math.BigDecimal;
 
-public record DefaultDaySetting(Integer dayOfWeek, BigDecimal countHours) {
+public record DefaultDaySetting(Integer dayOfWeek, int countHours) {
 }
